Mod ro ignore .gitkeep in site dir

+++ software/util/buildsite.py
@@ -76,7 +76,8 @@ def clear():
         if os.path.isdir(OUTPUTDIR):
             for root, dirs, files in os.walk(OUTPUTDIR):
                 for f in files:
-                    os.unlink(os.path.join(root, f))
+                    if f != ".gitkeep":
+                        os.unlink(os.path.join(root, f))
                 for d in dirs:
                     shutil.rmtree(os.path.join(root, d))
